+++ protocols/libp2p-request-response/src/lib.rs
@@ -56,6 +56,8 @@
 //! advertised during inbound respectively outbound protocol negotiation
 //! on the substreams.
 
+#![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
+
 pub mod codec;
 pub mod handler;